[ADD] lib update -> malloc reduction

+++ include/my_math.h
@@ -20,7 +20,7 @@ int my_pow(int number, int power);
 int my_int_sqrt(int nb);
 int my_is_prime(int nb);
 void my_sort_int_array(int *tab, int size);
-int my_nbr_length(long long nb);
+int my_nbr_length(long int nb);
 long long my_ll_pow(long long number, int power);
