+++ lib/my/my_stdio/src/post_processor.c
@@ -52,8 +52,8 @@ char *post_processor_float(char *float_str, parse_state_t *state)
     if (temp != NULL) return temp;
     int sign_space = state->flags.show_sign ||
         state->flags.space_if_pos || float_str[0] == '-';
-    int new_size = (my_strlen(new_res) > state->width ?
-        my_strlen(new_res) : state->width) + sign_space;
+    int new_size = (my_strlen(new_res) > (size_t)state->width ?
+        my_strlen(new_res) : (size_t)state->width) + sign_space;
     char *padded_res = my_calloc('\0', sizeof(char) * (new_size + 1));
     int utils[3] = { sign_space, new_size, float_str[0] == '-' };
     apply_alignment(new_res, padded_res, utils, state);
@@ -64,10 +64,10 @@ char *post_processor_float(char *float_str, parse_state_t *state)
 
 char *post_processor_str(char *str, parse_state_t *state)
 {
-    int new_size = my_strlen(str) > state->width ?
-        my_strlen(str) : state->width;
+    int new_size = my_strlen(str) > (size_t)state->width ?
+        my_strlen(str) : (size_t)state->width;
     if (state->width == 0 && state->precision > 0 &&
-        state->precision < my_strlen(str))
+        (size_t)state->precision < my_strlen(str))
         new_size = state->precision;
     char *new_res = my_calloc(' ', sizeof(char) * (new_size + 1));
     new_res[new_size] = '\0';
@@ -94,8 +94,8 @@ char *post_processor_int(char *res, parse_state_t *state)
     res = apply_precision(res, state, is_neg);
     int sign_space = state->flags.show_sign ||
         state->flags.space_if_pos || is_neg;
-    int new_size = (my_strlen(res) > state->width ?
-        my_strlen(res) : state->width) +
+    int new_size = (my_strlen(res) > (size_t)state->width ?
+        my_strlen(res) : (size_t)state->width) +
         (!state->flags.align_left && sign_space);
     char *new_res = my_calloc('\0', sizeof(char) *
         (new_size + sign_space + 1));
@@ -109,8 +109,8 @@ char *post_processor_uint(char *res, parse_state_t *state)
 {
     res = apply_precision(res, state, 0);
 
-    int new_size = (my_strlen(res) > state->width ?
-        my_strlen(res) : state->width);
+    size_t new_size = (my_strlen(res) > (size_t)state->width ?
+        my_strlen(res) : (size_t)state->width);
 
     char *new_res = my_calloc('\0', sizeof(char) *
         (new_size + 1 + (state->flags.hashtag ? 2 : 0)));
