+++ lib/my/my_str/my_strcpy.c
@@ -5,9 +5,11 @@
 ** copies a string into an other
 */
 
+#include "my_stdlib.h"
+
 void my_strcpy(char *dest, char const *src)
 {
-    int i = 0;
+    size_t i = 0;
 
     for (i = 0; src[i] != '\0'; i++)
         dest[i] = src[i];
