+++ lib/my/my_vec/misc.c
@@ -7,9 +7,9 @@
 
 #include "my_vec.h"
 
-int get_padded_size(int size)
+size_t get_padded_size(size_t size)
 {
-    int size_pad = 1;
+    size_t size_pad = 1;
 
     while (size_pad < size)
         size_pad *= 2;
@@ -17,10 +17,10 @@ int get_padded_size(int size)
     return size_pad;
 }
 
-void vec_swap(vec_t *v, int i, int j)
+void vec_swap(vec_t *v, size_t i, size_t j)
 {
     static char temp[MAX_VEC_ELEM_SIZE] = {0};
-    int el_size = v->base.el_size;
+    size_t el_size = v->base.el_size;
 
     my_memcpy(temp, v->data + (el_size * i), el_size);
     my_memcpy(v->data + (el_size * i), v->data + (el_size * j), el_size);
